backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    "Kacper Sylwar <75886014+KacperSylwar@users.noreply.github.com>"
]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<4.0"
django = ">=5.2.1,<6.0.0"
djangorestframework = ">=3.16.0,<4.0.0"
django-celery-beat = "^2.8.1"
django-celery-results = "*"
redis = "*"
celery = "*"
nordpool = "^0.4.5"
zeep = "^4.3.1"
requests = "^2.32.3"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"